+++ example/notification-popups/notificationPopups.js
@@ -1,6 +1,5 @@
-import Notifications from 'resource:///com/github/Aylur/ags/service/notifications.js';
-import Widget from 'resource:///com/github/Aylur/ags/widget.js';
-import { lookUpIcon } from 'resource:///com/github/Aylur/ags/utils.js';
+const notifications = await Service.import('notifications');
+const popups = notifications.bind('popups');
 
 /** @param {import('resource:///com/github/Aylur/ags/service/notifications.js').Notification} n */
 const NotificationIcon = ({ app_entry, app_icon, image }) => {
@@ -16,10 +15,10 @@ const NotificationIcon = ({ app_entry, app_icon, image }) => {
     }
 
     let icon = 'dialog-information-symbolic';
-    if (lookUpIcon(app_icon))
+    if (Utils.lookUpIcon(app_icon))
         icon = app_icon;
 
-    if (app_entry && lookUpIcon(app_entry))
+    if (app_entry && Utils.lookUpIcon(app_entry))
         icon = app_entry;
 
     return Widget.Icon(icon);
@@ -95,8 +94,6 @@ export const notificationPopup = Widget.Window({
     child: Widget.Box({
         class_name: 'notifications',
         vertical: true,
-        children: Notifications.bind('popups').transform(popups => {
-            return popups.map(Notification);
-        }),
+        children: popups.as(popups => popups.map(Notification)),
     }),
 });