+++ src/app.ts
@@ -38,14 +38,14 @@ export interface Config<W extends Gtk.Window = Gtk.Window> {
     icons?: string
     onWindowToggled?: (windowName: string, visible: boolean) => void
     onConfigParsed?: (app: App) => void
-    closeWindowDelay: { [key: string]: number }
+    closeWindowDelay?: { [key: string]: number }
 
     // FIXME: deprecated
-    notificationPopupTimeout: number
-    notificationForceTimeout: boolean
-    cacheNotificationActions: boolean
-    cacheCoverArt: boolean
-    maxStreamVolume: number
+    notificationPopupTimeout?: number
+    notificationForceTimeout?: boolean
+    cacheNotificationActions?: boolean
+    cacheCoverArt?: boolean
+    maxStreamVolume?: number
 }
 
 export class App extends Gtk.Application {
@@ -56,19 +56,23 @@ export class App extends Gtk.Application {
         });
     }
 
+
     private _dbus!: Gio.DBusExportedObject;
-    private _closeDelay!: Config['closeWindowDelay'];
     private _cssProviders: Gtk.CssProvider[] = [];
     private _objectPath!: string;
     private _windows: Map<string, Gtk.Window> = new Map();
     private _configPath!: string;
     private _configDir!: string;
 
+    private _closeWindowDelay!: Config['closeWindowDelay'];
+    get closeWindowDelay() { return this._closeWindowDelay || {}; }
+    set closeWindowDelay(v) { this._closeWindowDelay = v; }
+
     get windows() { return [...this._windows.values()]; }
     get configPath() { return this._configPath; }
     get configDir() { return this._configDir; }
 
-    resetCss() {
+    readonly resetCss = () => {
         const screen = Gdk.Screen.get_default();
         if (!screen) {
             console.error("couldn't get screen");
@@ -80,9 +84,9 @@ export class App extends Gtk.Application {
         });
 
         this._cssProviders = [];
-    }
+    };
 
-    applyCss(path: string) {
+    readonly applyCss = (path: string) => {
         const screen = Gdk.Screen.get_default();
         if (!screen) {
             console.error("couldn't get screen");
@@ -109,7 +113,7 @@ export class App extends Gtk.Application {
         );
 
         this._cssProviders.push(cssProvider);
-    }
+    };
 
     setup(bus: string, path: string, configDir: string, entry: string) {
         this.application_id = bus;
@@ -135,28 +139,28 @@ export class App extends Gtk.Application {
         this._load();
     }
 
-    connect(signal = 'window-toggled', callback: (_: this, ...args: any[]) => void): number {
+    readonly connect = (signal = 'window-toggled', callback: (_: this, ...args: any[]) => void) => {
         return super.connect(signal, callback);
-    }
+    };
 
-    toggleWindow(name: string) {
+    readonly toggleWindow = (name: string) => {
         const w = this.getWindow(name);
         if (w)
             w.visible ? this.closeWindow(name) : this.openWindow(name);
         else
             return 'There is no window named ' + name;
-    }
+    };
 
-    openWindow(name: string) {
+    readonly openWindow = (name: string) => {
         this.getWindow(name)?.show();
-    }
+    };
 
-    closeWindow(name: string) {
+    readonly closeWindow = (name: string) => {
         const w = this.getWindow(name);
         if (!w || !w.visible)
             return;
 
-        const delay = this._closeDelay[name];
+        const delay = this.closeWindowDelay[name];
         if (delay && w.visible) {
             timeout(delay, () => w.hide());
             this.emit('window-toggled', name, false);
@@ -164,17 +168,17 @@ export class App extends Gtk.Application {
         else {
             w.hide();
         }
-    }
+    };
 
-    getWindow(name: string) {
+    readonly getWindow = (name: string) => {
         const w = this._windows.get(name);
         if (!w)
             console.error(Error(`There is no window named ${name}`));
 
         return w;
-    }
+    };
 
-    removeWindow(w: Gtk.Window | string) {
+    readonly removeWindow = (w: Gtk.Window | string) => {
         const name = typeof w === 'string' ? w : w.name || 'gtk-layer-shell';
 
         const win = this._windows.get(name);
@@ -185,9 +189,9 @@ export class App extends Gtk.Application {
 
         win.destroy();
         this._windows.delete(name);
-    }
+    };
 
-    addWindow(w: Gtk.Window) {
+    readonly addWindow = (w: Gtk.Window) => {
         if (!(w instanceof Gtk.Window)) {
             return console.error(Error(`${w} is not an instanceof Gtk.Window, ` +
                 ` but it is of type ${typeof w}`));
@@ -206,7 +210,9 @@ export class App extends Gtk.Application {
         }
 
         this._windows.set(w.name, w);
-    }
+    };
+
+    readonly quit = () => super.quit();
 
     private async _load() {
         try {
@@ -220,7 +226,7 @@ export class App extends Gtk.Application {
             // FIXME:
             deprecated(config);
 
-            this._closeDelay = config?.closeWindowDelay || {};
+            this.closeWindowDelay = config?.closeWindowDelay || {};
 
             if (config.style) {
                 this.applyCss(config.style.startsWith('.')
@@ -228,9 +234,12 @@ export class App extends Gtk.Application {
                     : config.style);
             }
 
-            if (config.icons)
-                Gtk.IconTheme.get_default().append_search_path(config.icons);
-
+            if (config.icons) {
+                Gtk.IconTheme.get_default().append_search_path(
+                    config.icons.startsWith('.')
+                        ? `${this.configDir}${config.icons.slice(1)}`
+                        : config.icons);
+            }
             if (typeof config.onWindowToggled === 'function')
                 this.connect('window-toggled', (_, n, v) => config.onWindowToggled!(n, v));