+++ src/service/network.ts
@@ -90,12 +90,12 @@ export class Wifi extends Service {
         }
     }
 
-    scan() {
+    readonly scan = () => {
         this._device.request_scan_async(null, (device, res) => {
             device.request_scan_finish(res);
             this.emit('changed');
         });
-    }
+    };
 
     private _activeAp() {
         if (this._ap)
@@ -242,9 +242,9 @@ export class Network extends Service {
         }
     }
 
-    toggleWifi() {
+    readonly toggleWifi = () => {
         this._client.wireless_enabled = !this._client.wireless_enabled;
-    }
+    };
 
     private _getDevice(devType: NM.DeviceType) {
         return this._client