+++ src/utils.ts
@@ -6,6 +6,8 @@ import * as Timeout from './utils/timeout.js';
 import * as Fetch from './utils/fetch.js';
 import * as Notify from './utils/notify.js';
 import * as Pam from './utils/pam.js';
+import * as Gobject from './utils/gobject.js';
+import * as Binding from './utils/binding.js';
 
 export const USER = GLib.get_user_name();
 export const HOME = GLib.get_home_dir();
@@ -21,6 +23,7 @@ export const {
     readFile,
     readFileAsync,
     writeFile,
+    writeFileSync,
     monitorFile,
 } = File;
 
@@ -46,6 +49,17 @@ export const {
 export const { fetch } = Fetch;
 export const { notify } = Notify;
 
+export const {
+    kebabify,
+    pspec,
+    registerGObject,
+} = Gobject;
+
+export const {
+    merge,
+    derive,
+} = Binding;
+
 export default {
     USER,
     HOME,
@@ -58,6 +72,7 @@ export default {
     readFile,
     readFileAsync,
     writeFile,
+    writeFileSync,
     monitorFile,
 
     timeout,
@@ -75,4 +90,11 @@ export default {
 
     authenticate,
     authenticateUser,
+
+    kebabify,
+    pspec,
+    registerGObject,
+
+    merge,
+    derive,
 };