+++ src/widgets/circularprogress.ts
@@ -12,7 +12,7 @@ interface Context {
 }
 
 export type CircularProgressProps<
-    Child extends Gtk.Widget,
+    Child extends Gtk.Widget = Gtk.Widget,
     Attr = unknown,
     Self = CircularProgress<Child, Attr>
 > = BaseProps<Self, Gtk.Bin.ConstructorProperties & {
@@ -52,7 +52,6 @@ export class CircularProgress<
 
     get child() { return super.child as Child; }
     set child(child: Child) { super.child = child; }
-
 
     get rounded() { return this._get('rounded') || false; }
     set rounded(r: boolean) {