+++ src/widgets/eventbox.ts
@@ -5,7 +5,7 @@ import Gdk from 'gi://Gdk?version=3.0';
 type EventHandler<Self> = (self: Self, event: Gdk.Event) => boolean | unknown;
 
 export type EventBoxProps<
-    Child extends Gtk.Widget,
+    Child extends Gtk.Widget = Gtk.Widget,
     Attr = unknown,
     Self = EventBox<Child, Attr>,
 > = BaseProps<Self, Gtk.EventBox.ConstructorProperties & {
@@ -107,7 +107,6 @@ export class EventBox<Child extends Gtk.Widget, Attr> extends Gtk.EventBox {
 
     get child() { return super.child as Child; }
     set child(child: Child) { super.child = child; }
-
 
     get on_hover() { return this._get('on-hover'); }
     set on_hover(callback: EventHandler<this>) {