+++ src/widgets/widget.ts
@@ -94,8 +94,8 @@ export type BaseProps<Self, Props, Attr = unknown> = {
     setup?: (self: Self) => void
 } & BindableProps<CtorProps<Props & CommonProps<Attr>>>
 
-type Requierd<T> = { [K in keyof T]-?: T[K] };
-export interface Widget<Attr> extends Requierd<CommonProps<Attr>> {
+type Required<T> = { [K in keyof T]-?: T[K] };
+export interface Widget<Attr> extends Required<CommonProps<Attr>> {
     hook<
         Gobject extends GObject.Object,
     >(
@@ -215,7 +215,6 @@ export class AgsWidget<Attr> extends Gtk.Widget implements Widget<Attr> {
         timeout: number,
         callback: (self: this) => void,
     ): this {
-        callback(this);
         interval(timeout, () => callback(this), this);
         return this;
     }
@@ -259,7 +258,7 @@ export class AgsWidget<Attr> extends Gtk.Widget implements Widget<Attr> {
         this.connect('destroy', () => this._set('is-destroyed', true));
 
         idle(() => {
-            if (this.click_through)
+            if (this.click_through && !this.is_destroyed)
                 this.input_shape_combine_region(new Cairo.Region);
         });
 
@@ -431,33 +430,6 @@ export class AgsWidget<Attr> extends Gtk.Widget implements Widget<Attr> {
             return;
 
         this.setCss(css);
-    }
-
-    get child(): Gtk.Widget | null {
-        // @ts-expect-error
-        if (typeof this.get_child === 'function')
-            // @ts-expect-error
-            return this.get_child();
-
-        return null;
-    }
-
-    set child(child: Gtk.Widget) {
-        if (this.child !== child && this.child)
-            this.child.destroy();
-
-        // @ts-expect-error
-        if (typeof this.set_child === 'function')
-            // @ts-expect-error
-            this.set_child(child);
-
-        // @ts-expect-error
-        else if (typeof this.add === 'function')
-            // @ts-expect-error
-            this.add(child);
-
-        else
-            console.error(Error(`can't set child on ${this}`));
     }
 
     _updateCursor() {